Rename utility class.
